chore(useAppMetrics, Components): reorg hooks and use proper keys
