+++ packages/frontend/src/known-query-suffixes.js
@@ -1,0 +1,12 @@
+export const KNOWN_QUERY_SUFFIXES = {
+  DAILY_BREAKDOWN_METRICS: 'DAILY_BREAKDOWN_METRICS',
+  HOURLY_LATENCY_METRICS: 'HOURLY_LATENCY_METRICS',
+  LATEST_RELAYS: 'LATEST_RELAYS',
+  ONCHAIN_DATA: 'ONCHAIN_DATA',
+  PREVIOUS_SUCCESSFUL_METRICS: 'PREVIOUS_SUCCESSFUL_METRICS',
+  PREVIOUS_TOTAL_METRICS: 'PREVIOUS_TOTAL_METRICS',
+  SESSION_METRICS: 'SESSION_METRICS',
+  USER_APPS: 'USER_APPS',
+  WEEKLY_SUCCESSFUL_METRICS: 'WEEKLY_SUCCESSFUL_METRICS',
+  WEEKLY_TOTAL_METRICS: 'WEEKLY_TOTAL_METRICS',
+}
